Add intent specification compiler.
The runtime will load an intent specification from a file called
<INTENT_LABEL>.intent, where <INTENT_LABEL> is the
label passed as the first parameter of the optimize() routine.

The runtime will similarly load the knob and measure models
from CSV files called <INTENT_LABEL>.knobtable and
<INTENT_LABEL>.measuretable. These files contain one
column per knob and measure listed in the intent specification,
respectively, as well as a common column 'id', by which the
two files are joined in the runtime.

This relates to PRO-61.