Add support for exhaustive profiling of FAST programs.
To compute an exhaustive profile, run `make profile`. This will
profile each configuration specified in the intent file as an
average over 1000 inputs. To change the number of inputs, run
`make size=<N> profile`, where `<N>` is the desired number.

NOTE: The profile will be incrementally written to two files
`<NAME>.knobtable` and `<NAME>.measuretable`, where `<NAME>`
is the name of the application, if such a name has been specified
by constructing an Application object, or else `<NAME>` is "fast".