Make incrementer emulatable.
Add a targets `emulate` and `emulate-scripted` to the Makefile,
that start the incrementer in emulated interaction mode.

Note that running in scripted mode will make the main application
thread block until a process request has been received. The command
line output that is produced between process requests comes from
the MeasuringDevice's TimingSamplingPolicy, that records values of
measures on a regular basis (every 100ms).