Add fetching of initialization parameters from brass-th/ready.
When the environment variable "proteus_runtime_initializeFromTestHarness"
is set to "true" then, once the FAST REST API is up (confirmed by using the
RestClient to post to http://127.0.0.1/alive), a post is performed to
http://<TEST-HARNESS-HOST>:<TEST-HARNESS-PORT>/ready, where
<TEST-HARNESS-HOST> is read from the environment variable
"proteus_client_rest_serverPath" and <TEST-HARNESS-PORT> is read
is read from the environment variable "proteus_client_rest_serverPort".
(by default, it is "80") on a port that is read from 80.
After the initialization parameters have been parsed, a post to
http://<TEST-HARNESS-HOST>:<TEST-HARNESS-PORT>/initialized
is performed.

In this commit, the only part of the initialization parameters received from
the test harness that is used is the missionIntent.

Because the FastRestServer is currently started from the optimize() function
(as a quick-and-dirty workaround to the lazy initialization of static variables,
which prevented the server from starting if spawned during the initialization
of a static variable in the Runtime), the tests for this functionality currently
reside in OptimizeTests.swift.