Eliminate Runtime overhead from measure computations.
Both time (latency, performance) and energy (energyDelta, powerConsumption)
are now computed so that only resources spent processing during the execution
of the optimize loop body are measured, that is, the overhead incurred by runtime
operations such as schedule computation and reconfiguration is not included.

Also, only reconfigure the system (by applying knob settings) when the configuration
has changed since the last iteration. This is to avoid the overhead of costly
reconfiguration, such as setting system configuration knobs.