Proof-of-concept of sceneImportance test parameter.
This parameter replaces sceneObfuscation with a simpler concept
that exhibits a similar reaction by the system. When sceneImportance
is high, the system must ensure that the quality (measure) of the
output is high. The sceneImportance parameter thus provides a
way to filter the configurations that are considered by the controller,
to leave only those that produce a sufficiently high quality, relative
to the maximum possible one.