Introduce enforceEnergyLimit test parameter.
This toggles the following semantics for the missionLength parameter,
which is always used to determine the number of iterations the
application should execute.

Based on the current model, the configuration with maximal energyDelta
is identified. This energyDelta is multiplied by missionLength as an
estimated upper bound "energyLimit" on the energy required to complete
the initial missionLength. If the missionLength is changed during execution,
this energyLimit is used to deduce which configurations use too much energy
and should be filtered out to meet the new missionLength.