Use filtered model when enforcing energyLimit.
If the controller model is APP.{knob,measure}table, and when the
enforceEnergyLimit initialization parameter is true, then the
model APP.{knob,measure}table.filtered will be used instead.
The latter is assumed to be a subset of the former model, and
contain only configurations such that quality is an isotonic
function of energyDelta.