Enable dynamic initialization of Knob objects.
With this change, Knobs initialized inside the optimize loop body
are automatically registered as application knobs.
This is in contrast to knobs created outside the optimize body,
which are created before an Runtime instance is available, and
need to be passed to the optimize construct to be registered
in the runtime after it is initialized.