Add support for restricting a knob's values at runtime.
Using two new public methods in the Knob type, restrict() and control(),
the range of values that a knob can be set to by the controller can be
restricted (to an array of values passed to restrict()) and unrestricted,
respectively.

The utility methods Knob.constant() can be used to restrict the knob
to a single value. If no parameter is passed to either restrict() or
constant(), the values of the knob will be restricted to the current value
of the knob.

The control() method can be used to "undo" restrictions applied using
the restrict() and constant() methods, effectively allowing the controller
to assign to the knob any value listed in the intent specification.

The new methods are implemented using Runnable.setApplicationKnobFilter,
which relies on a new mechanism for filtering the model used by the controller.
When Knob.restrict(vs) is called, a "filter" is added to the Runtime.modelFilters
dictionary, and then the controller is reinitialized. During initialization, the filters
are "applied", so that only configurations whose values for this knob are
included in the "vs" parameter passed remain in the model.