Remove enforceEnergyLimit and sceneImportance test params.
When these parameters were set (enforceEnergyLimit to true, and sceneImportance
to a value greater than 0.0 and less than or equal to 1.0), the model used to control
the application was filtered according to each parameter.

Basically, when enforceEnergyLimit was enabled, configurations that used too much
energy per iteration, that is, exceeded the energy limit per iteration that was computed
at the start (based on the assumption that the most energy-consuming configuration
was going to be used throughout the entire execution), were eliminated from the model,
based on the current missionLength.

When sceneImportance was enabled, configurations that has a too-low quality were
(relative to the maximum quality achievable while meeting the current intent) were
eliminated from the model.

With this commit, the above behavior is no longer implemented by the runtime.