Make collection / reporting of detailed statistics optional.
Per-configuration statistics take a large part of the execution time
of a typical x264 iteration, and thus affect the ability to effectively
profile and control that application. Thus, the gathering of these
statistics needs to be optimized. Until this is done, this functionality
is off by default, and can be turned on again by setting the following
new environment variable to "true":

	proteus_runtime_collectDetailedStats