Work-around to parse until knob constraints.
Until this can be handled by swift-ast. Uses regular expressions to extract knob constraint
from intent spec, and then create a version of the intent spec without the knob constraint,
to be parsed by the existing parser. Then, use NSPredicate to implement evaluation of
the knob constraint in the new satisfiesKnobConstraints(knobSettings: KnobSettings)
method of IntentSpec.