Unified model filtering and feedback bootstrapping.
Now the trimming of models w.r.t application knob restrictions
and intent specification restrictions (due to passing to the /perturb
end-point either an intent with a smaller knob range, or passing
scenario knobs availableCores or availableCoreFrequency ) will
register a predicate in Runtime.modelFilters, as was already done
following a call to Runtime.restrict().

Then, whenever Runtime.registerModel() [previously Runtime.setModel()]
is called, these filters are automatically applied to the model.
Unless overridden, the first configuration in this model is then picked out
of the filtered model, and a ConstantController is then instantiated and
activated in the runtime. The, Optimize.run() runs the application until at
least one window has passed, to ensure that the measure values of the
picked-out configuration have stabilized, and these values are then used
to initialize the adaptive controller.