Introduce WarmupController.
This is a wrapper for the adaptive controllers that enforces the following.
After a perturbation occurs, that is, after: 1) a call to the /perturb REST
end-point changes the intent or a scenario knob (availableCoreFrequency
or availableCores); 2) a call to the Knob.restrict() API changes the range
of an application knob; then a WarmupController is initialized.

This controller takes an initial configuration as a parameter, as well as
another instance of the Controller class. When the WarmupController's
getSchedule() method is called for the first time, it will return a schedule
based on the initial configuration, so that the system runs in a fixed
configuration during the initial window after the perturbation.
This ensures that the measure statistics that are gathered for this (known)
configuration are accurate.
All subsequent calls to the WarmupController's getSchedule() method
result in a forward call to its wrapped controller.