Delay resetting controller until next iteration.
When registering a perturbation or changing intent using the
Runtime.registerPerturbation and Runtime.changeIntent methods,
do not directly call the Runtime.registerIntentAndModel method.
Instead, set the Runtime.perturbationOccurred flag to tru, which
will cause a call to Runtime.registerIntentAndModel to happen in
the optimize function at the beginning of the next iteration.
This will then apply any changed intent, scenario knobs and knob
restriction model filters.